Windows: fix "invalid drive" errors on Windows 7
We can't open a handle on an ejected drive letter on Windows 7, which
means that our `EjectDriveLetter` safety retry look causes the program
to fail with an error even though the drive was unmounted correctly.

As a solution, we abort the loop if previous tries succeeded and we get
`MOUNTUTILS_ERROR_INVALID_DRIVE` (which means we couldn't open a handle
on the drive).

This error only happens on Windows 7, and I only experienced it on
internal SD Card readers and external HDDs.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>