Fix "Unmount failed" error on Windows
We've been recently seeing tons of "Unmount failed" errors coming from
Windows. After some inspection, all the debugging stack traces show the
following scenario:

- X drive letters are scanned, but none belongs to the drive we're
  trying to unmount
- Finally, one of the drive letter matches
- The matching drive letter is successfully unmounted
- We continue scanning more drive letters
- We get an invalid handle error when opening such drive letter

This issue happens when the computer mounts network drive letters (which
get usually mounted using letters like X, and Z). The code sees that
such as drive letter exists, and tries to open a handle for it, which
fails with an invalid handle error.

There is another edge case where we've seen this bug.

Note that we get a list of logical drives once mountutils start, and we
keep it on memory in order to refer to it as we traverse through the
available drive letters.

If we scan than drive F: is available for example, we don't have any
guarantee that such drive letter is still available by the time we're
interacting with it (which is made worse by the fact we have retry
timeouts in some parts of the code).

This issue can be reproduced by plugging two drives, ejecting the drive
letter that comes first, and quickly removing the other drive while the
selected drive is being unmounted.

As a solution, we re-check that the drive letter is valid every time
we're about to access it.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>