fix(macos): retry the unmount operation if loop times out
We've been experiencing a large amount of "Unmount failed" errors on
v1.1.0, all of them caused by the unmount wait loop timing out, without
any unmount success result.

The best solution I could find, is simply retrying the whole unmount
operation if the loop times out, which required adding a new error flag
to track loops timing out, and an extra number argument to keep track of
the amount of times we retry (in order to not retry forever).

Reproducing the issue is a bit tricky, but I found out the following
steps allow the error to happen around 50% of times on my system:

- Cause a lot of disk activity by creating a script that executes
  `diskutil list` every 500ms
- Reformat the drive to FAT32 with MBR using Disk Utility
- Attempt to flash an image on the Etcher tool
- Retry as needed until the error occurs

Change-Type: patch
Changelog-Entry: Retry the unmount operation if the loop times out in macOS.
Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>